[Bug 4590] Leading zero added to TimeChooser display in Internet Explorer
Wow. We know that String.substr() is not in the ECMAscript standard. Who would
of thought, though, that IE8 would stray so far from the established
implementations of that function. Apparently, IE8 does not take
String.substr(-2) to mean the last two characters of the string. Sigh.

I shouldn't have used substr() anyway, since it is not
ECMAscript-compliant. This patch replaces the uses of substr() with
substring() and has been confirmed by Andy Jarrell to work as intended.

Derrell